Add the ability to view who's signed up for a topic.
Add the ability of the owner of a topic to view the list of users who
have signed up for his/her topic. Later on, I'll add the ability to
specify whether a topic is "anonymous signup" (i.e. owner doesn't see
who's signed up, only a count of those signed up).